+++ x/currency/msg.go
@@ -4,7 +4,12 @@ import (
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/coin"
 	"github.com/iov-one/weave/errors"
+	"github.com/iov-one/weave/migration"
 )
+
+func init() {
+	migration.MustRegister(1, &NewTokenInfoMsg{}, migration.NoModification)
+}
 
 var _ weave.Msg = (*NewTokenInfoMsg)(nil)
 
@@ -13,6 +18,9 @@ func (NewTokenInfoMsg) Path() string {
 }
 
 func (t *NewTokenInfoMsg) Validate() error {
+	if err := t.Metadata.Validate(); err != nil {
+		return errors.Wrap(err, "metadata")
+	}
 	if !coin.IsCC(t.Ticker) {
 		return errors.Wrapf(errors.ErrCurrency, "invalid ticker: %s", t.Ticker)
 	}